Fixed dependency issue with policy
