+++ Board.jl
@@ -1,4 +1,4 @@
-struct Board
+mutable struct Board
     bounds::Vector
     layout::Matrix
 end
